+++ py/appbase/database.py
@@ -6,6 +6,7 @@ db = SQLAlchemy()
 
 def init_db(app):
     db.init_app(app)
+    migrate = Migrate(app, db)
 
     # modelとして定義した全テーブルをCreateする
     db.create_all(app=app)
